Add undo support to link command
Adding a `-u` flag for the link command which will undo the symlinking
process. It works by deleting all symlinks that point to a file in the
source directory.

For example, if you created a file with the matching path, but it wasn't
symlinked to the source directory, it would not be deleted by running
`link -u`.